[BTRX-734-735-736] Data use Letter Link, PDF Box Update and Alert Time extension (#211)
* refactor code

* style changes and spinner fix

* Error alert fix

* spacing

* Move DUL to component
