+++ build.gradle
@@ -121,7 +121,7 @@ dependencies {
     compile group: 'org.apache.httpcomponents', name: 'httpmime', version: '4.5.6'
     compile 'org.grails.plugins:gorm-logical-delete:2.0.0.M2'
     // https://mvnrepository.com/artifact/org.apache.pdfbox/pdfbox
-    compile group: 'org.apache.pdfbox', name: 'pdfbox', version: '2.0.13'
+    compile group: 'org.apache.pdfbox', name: 'pdfbox', version: '2.0.15'
 
     testCompile ('com.github.tomakehurst:wiremock:2.18.0') {
         exclude group: 'com.fasterxml.jackson.core', module: 'jackson-databind'
