+++ src/main/webapp/projectContainer/ConsentGroups.js
@@ -116,12 +116,13 @@ export const ConsentGroups = hh(class ConsentGroups extends Component {
         return prev;
       });
     });
-    // Display for 8 seconds a message indicating the submission of a new consent group. This is temporary until this page is moved to react.
+    // Display for 8 seconds a message indicating the submission of a new consent group. 
+    // This is temporary until this page is moved to react.
     // https://broadinstitute.atlassian.net/browse/BTRX-628
     var url = new URLSearchParams(window.location.search);
     if (url.get('tab') === 'consent-groups' && url.has('new')) {
       $('#alert').fadeIn('slow', function () {
-        $('#alert').delay(8000).fadeOut();
+        $('#alert').delay(15000).fadeOut();
         history.pushState({}, null, window.location.href.split('&')[0]);
       });
     }
